# src/main/java/com/proyecto/challengejava/dto/AuthResponse.java
package com.proyecto.challengejava.dto;

public class AuthResponse {
    private String token;

    public AuthResponse() {}

    public AuthResponse(String token) {
        this.token = token;
    }

    public String getToken() {
        return token;
    }

    public void setToken(String token) {
        this.token = token;
    }
}